don't encode the data qsarg as JSON
the JS SDK encodes data as JSON because its goal is to put the data dictionary into a GET qsarg. it does this by writing data as urlencoded JSON.
in the mobile SDK, our goals is to put data into a POST request body that is itself JSON encoded. There's no need to do an extra JSON encode on data
since the POST body will handle a single big JSON blob just fine