* don't encode the data qsarg as JSON
the JS SDK encodes data as JSON because its goal is to put the data dictionary into a GET qsarg. it does this by writing data as urlencoded JSON.
in the mobile SDK, our goal is to put data into a POST request body that is itself JSON encoded. There's no need to do an extra JSON encode on data
since the JSON post body can handle a nested dictionary just fine

* pass the current time not as "rand" but as "data.ts", or what putils refers to as the "override timestamp"