Relax `testBackoff()` interval equality to a 0.04 accuracy
We've seen this test being flaky and failing when the tracked interval
was a few hundredth of a second different from the expected 13.65.
See for example
https://github.com/Automattic/AnalyticsSDK-iOS/pull/6#issuecomment-1508327314

It seems in the context of a backoff implementation we can account for a
few hundredth of a second difference between the expected interval and
the recorded one.

This approach look even more acceptable when we consider that
`testSampleFn()` also has an accuracy range of 0.2 seconds, although
implemented in a different way. See
https://github.com/Automattic/AnalyticsSDK-iOS/blob/6f0ab679f0de687e09814d75acac7aee75e48e96/ParselyTrackerTests/SamplerTests.swift#L33-L48

For what is worth, I tried updating the underlying scheduling code to
use a finer grained `Stride` `delay`, both with `.milliseconds` and
`.microseconds` and the test was still flaky (measured as failed
attempts over 10 repetitions).